Some added verbosity in the JSyphonServer to help debug implementations, and a pure Java method to call ( test() ) to ensure linking to the Jar works, when troubleshooting JNI / library loading, etc. Also cleaned up the ServerTest just a touch.
We ought to add a namespace to the top level Jar so we import as v002.syphon. I am also tempted to do away with the JSyphon naming for classes, as its *clear* we are in Java land when using Java. JSyphon only really needs to be named so its clear what it is in our SVN... dont we think - I see no reason to import jsyphon.* vs v002.syphon.* - its cleaner. I think. Will attempt to move to that later. Need to focus on actual features first ;)